A few conversions from float to int and defining constants as float (rather than double) to reduce warnings with Intel comiler.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@2125 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/columns/Communicator.cpp
@@ -20,7 +20,7 @@ Communicator::Communicator(int* argc, char*** argv)
 
    sprintf(commName, "[%2d]: ", icRank);
 
-   r = sqrt(worldSize);
+   r = sqrtf(worldSize);
    numRows = (int) r;
    numCols = (int) worldSize / numRows;
